+++ GWToolbox/GWToolbox/ToolboxModule.h
@@ -19,8 +19,8 @@ public:
 	// Terminate module
 	virtual void Terminate() {};
 
-	// Update. Will always be called once every frame.
-	virtual void Update(DWORD delta) {};
+	// Update. Will always be called once every frame. Delta in seconds
+	virtual void Update(float delta) {};
 
 	// This is provided (and called), but use ImGui::GetIO() during update/render if possible.
 	virtual bool WndProc(UINT Message, WPARAM wParam, LPARAM lParam) { return false; };
